backend/requirements.txt
# Base dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.21.1  # For production-ready server
pydantic>=2.0.0
pyyaml>=6.0

# Database
sqlalchemy>=2.0.0
databases>=0.5.0
asyncpg>=0.27.0  # PostgreSQL driver

# Security
passlib>=1.7.4
python-jose>=3.3.0

# Utilities
python-dotenv>=1.0.0

# AI and Web related
openai>=1.0.0
requests>=2.31.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=1.5.0
openpyxl>=3.0.0